Fix problem with newlines in Thunderbird due to some keymappings not mapping the \n keysym. Fix crash on startup due to new 'safe' persistence code. Improvements to copy and paste functions.
git-svn-id: http://autokey.googlecode.com/svn/trunk@64 65b5f37c-86d1-11de-b610-71a7c637a048

+++ src/lib/configurationmanager.py
@@ -55,8 +55,9 @@ def save_config(configManager):
     configManager.configHotkey.set_closure(None)
     configManager.toggleServiceHotkey.set_closure(None)
     configManager.showPopupHotkey.set_closure(None)
-    # Back up configuration
-    shutil.copy(CONFIG_FILE, CONFIG_FILE_BACKUP)
+    # Back up configuration if it exists
+    if os.path.exists(CONFIG_FILE):
+        shutil.copy(CONFIG_FILE, CONFIG_FILE_BACKUP)
     try:
         outFile = open(CONFIG_FILE, "wb")
         pickle.dump([ConfigurationManager.SETTINGS, configManager], outFile)
